+++ zulipterminal/config/keys.py
@@ -77,6 +77,11 @@ KEY_BINDINGS = {
         'keys': {'S'},
         'help_text': 'Narrow to the topic of the current message',
     },
+    'TOGGLE_NARROW': {
+        'keys': {'z'},
+        'help_text':
+            'Narrow to a topic/private-chat, or stream/all-private-messages',
+    },
     'ALL_PM': {
         'keys': {'P'},
         'help_text': 'Narrow to all private messages',